+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.2.0 / 2021-05-25
 - Updated crates
 - Updated gdnative to 0.9.3
-- Changed from using project.toml to godot-rust-cli.toml
+- Changed from using project.toml to godot-rust-cli.json
 
 ## 0.1.2 / 2021-02-11
 - Moved docs into their own branch.